Committing the unit tests written by Saleem to the master branch. Only some of them are working as of right now.
In order for my eclipse project to recognize the netbeans jUnit test
layout, the jUnit package declarations have been changed like so:

implTests --> UnitTests.ImplTests

This is because Eclipse sees all folders in the workspace as a package.
There's a small chance this might lead to buggy behavior in NetBeans.