Added another Exception class: NegativeElevatorException
Refined the messages that are printed out for each custom exception so
the value in question is printed out along with the message.

Added some additional documentation. This is not meant to be an
exhaustive change. The modified classes were modified for one of two
reasons:

1) To explicitly state whether an integer parameter value is meant to be
zero-based or one-based
2) While working with the class, I noticed that some documentation was
missing